+++ src/Application/Interfaces/IClothingService.cs
@@ -6,8 +6,8 @@ namespace Application.Interfaces
     {
         Task<IEnumerable<ClothingAccountDto>> GetAllUserProductByUserId(string id);
         Task CreateClothing(ClothingCreateDto clothing);
-        Task<int?> GetClothingPriceById(int clothingId);
-        Task ChangePriceById(int id, int price);
+        Task<decimal?> GetClothingPriceById(int clothingId);
+        Task ChangePriceById(int id, decimal price);
         Task<IEnumerable<ClothingItemPageDto>> GetAllWithPagination(ClothingPaginationFilterDto paginationFilterDto);
         Task<int> GetCount(int? typeId, int? sizeId);
     }
